refactor: member 패키지 분리
- auth 패키지와 구분하기 위해 member 패키지 미리 분리
