+++ src/test/java/com/flab/eattofit/member/infrastructure/member/MemberJpaRepositoryTest.java
@@ -1,6 +1,6 @@
-package com.flab.eattofit.member.infrastructure;
+package com.flab.eattofit.member.infrastructure.member;
 
-import com.flab.eattofit.member.domain.Member;
+import com.flab.eattofit.member.domain.member.Member;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
